Add a method of importing data
This attempt uses migrations to get the data in. It works, but it's not
ideal: the data shows up in the tests, where it's not expected, and
simply reruning `docker-compose up` will cause the database to no longer
contain the data, as migrations aren't automatically rerun. I'm going to
look into using psycopg2 directly. Then it'll just be a matter of
getting the tests passing again.