Add tag search, fix reversal lookup bug
Tag lookup is a bit scuffed -- it grabs all tags and manually filters
them. I had to change the url encoding for the channel view -- some
channels actually use '/' as part of their name, which breaks the search
if they come up in the results. That's fixed now, at the cost of more
complex urls.